+++ packages/vscode-extension/src/debugging/DebugAdapterDescriptorFactory.ts
@@ -1,11 +1,20 @@
 import vscode from "vscode";
 import { DebugAdapter } from "./DebugAdapter";
+import { CDPDebugAdapter } from "./CDPDebugAdapter";
 
 export class DebugAdapterDescriptorFactory implements vscode.DebugAdapterDescriptorFactory {
   createDebugAdapterDescriptor(
     session: vscode.DebugSession
   ): vscode.ProviderResult<vscode.DebugAdapterDescriptor> {
+    return new vscode.DebugAdapterInlineImplementation(new DebugAdapter(session));
+  }
+}
+
+export class CDPDebugAdapterDescriptorFactory implements vscode.DebugAdapterDescriptorFactory {
+  createDebugAdapterDescriptor(
+    session: vscode.DebugSession
+  ): vscode.ProviderResult<vscode.DebugAdapterDescriptor> {
     const configuration = session.configuration;
-    return new vscode.DebugAdapterInlineImplementation(new DebugAdapter(configuration));
+    return new vscode.DebugAdapterInlineImplementation(new CDPDebugAdapter(configuration));
   }
 }
