+++ packages/vscode-extension/src/extension.ts
@@ -15,7 +15,10 @@ import vscode from "vscode";
 import { TabPanel } from "./panels/Tabpanel";
 import { PreviewCodeLensProvider } from "./providers/PreviewCodeLensProvider";
 import { DebugConfigProvider } from "./providers/DebugConfigProvider";
-import { DebugAdapterDescriptorFactory } from "./debugging/DebugAdapterDescriptorFactory";
+import {
+  CDPDebugAdapterDescriptorFactory,
+  DebugAdapterDescriptorFactory,
+} from "./debugging/DebugAdapterDescriptorFactory";
 import { Logger, enableDevModeLogging } from "./Logger";
 import {
   extensionContext,
@@ -223,9 +226,23 @@ export async function activate(context: ExtensionContext) {
   );
 
   context.subscriptions.push(
+    debug.registerDebugConfigurationProvider(
+      "com.swmansion.js-debugger",
+      new DebugConfigProvider(),
+      DebugConfigurationProviderTriggerKind.Dynamic
+    )
+  );
+
+  context.subscriptions.push(
     debug.registerDebugAdapterDescriptorFactory(
       "com.swmansion.react-native-debugger",
       new DebugAdapterDescriptorFactory()
+    )
+  );
+  context.subscriptions.push(
+    debug.registerDebugAdapterDescriptorFactory(
+      "com.swmansion.js-debugger",
+      new CDPDebugAdapterDescriptorFactory()
     )
   );
